@@ -10,22 +10,58 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.2.0",
-    "@angular/common": "^14.2.0",
-    "@angular/compiler": "^14.2.0",
-    "@angular/core": "^14.2.0",
-    "@angular/forms": "^14.2.0",
-    "@angular/platform-browser": "^14.2.0",
-    "@angular/platform-browser-dynamic": "^14.2.0",
-    "@angular/router": "^14.2.0",
+    "@angular/animations": "^15.0.4",
+    "@angular/cdk": "^13.0.0",
+    "@angular/common": "^15.0.4",
+    "@angular/compiler": "^15.0.4",
+    "@angular/core": "^15.0.4",
+    "@angular/forms": "^15.0.4",
+    "@angular/localize": "^15.0.4",
+    "@angular/platform-browser": "^15.0.4",
+    "@angular/platform-browser-dynamic": "^15.0.4",
+    "@angular/router": "^15.0.4",
+    "@fortawesome/free-solid-svg-icons": "^6.2.1",
+    "@fullcalendar/angular": "^6.0.1",
+    "@fullcalendar/core": "^6.0.1",
+    "@fullcalendar/daygrid": "^6.0.1",
+    "@ng-bootstrap/ng-bootstrap": "^13.1.1",
+    "@popperjs/core": "^2.10.2",
+    "@progress/kendo-angular-buttons": "latest",
+    "@progress/kendo-angular-common": "latest",
+    "@progress/kendo-angular-dateinputs": "latest",
+    "@progress/kendo-angular-dialog": "latest",
+    "@progress/kendo-angular-dropdowns": "latest",
+    "@progress/kendo-angular-inputs": "latest",
+    "@progress/kendo-angular-intl": "latest",
+    "@progress/kendo-angular-l10n": "latest",
+    "@progress/kendo-angular-popup": "latest",
+    "@progress/kendo-angular-scheduler": "^4.3.7",
+    "@progress/kendo-angular-treeview": "latest",
+    "@progress/kendo-data-query": "latest",
+    "@progress/kendo-date-math": "latest",
+    "@progress/kendo-drawing": "latest",
+    "@progress/kendo-licensing": "latest",
+    "@progress/kendo-recurrence": "latest",
+    "@syncfusion/ej2-angular-schedule": "^20.4.38-ngcc",
+    "@syncfusion/ej2-grids": "^20.4.38",
+    "@syncfusion/ej2-material-theme": "~20.4.38",
+    "bootstrap": "^5.2.3",
+    "bootstrap-vue": "^2.23.1",
+    "json-server": "^0.17.1",
+    "ng2-order-pipe": "^0.1.5",
+    "ng2-search-filter": "^0.5.1",
+    "ngx-bootstrap": "^10.2.0",
+    "ngx-order-pipe": "^2.2.0",
+    "ngx-pagination": "^6.0.3",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
+    "vue": "^3.2.45",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^14.2.9",
-    "@angular/cli": "~14.2.9",
-    "@angular/compiler-cli": "^14.2.0",
+    "@angular-devkit/build-angular": "^15.0.4",
+    "@angular/cli": "~15.0.4",
+    "@angular/compiler-cli": "^15.0.4",
     "@types/jasmine": "~4.0.0",
     "jasmine-core": "~4.3.0",
     "karma": "~6.4.0",
@@ -33,6 +69,6 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.0.0",
-    "typescript": "~4.7.2"
+    "typescript": "~4.8.4"
   }
 }
