@@ -1,6 +1,6 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import {ComponentFixture, TestBed} from '@angular/core/testing';
 
-import { UserLoginComponent } from './user-login.component';
+import {UserLoginComponent} from './user-login.component';
 
 describe('UserLoginComponent', () => {
   let component: UserLoginComponent;
@@ -8,9 +8,9 @@ describe('UserLoginComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ UserLoginComponent ]
+      declarations: [UserLoginComponent]
     })
-    .compileComponents();
+      .compileComponents();
 
     fixture = TestBed.createComponent(UserLoginComponent);
     component = fixture.componentInstance;
