@@ -1,7 +1,7 @@
-import { User } from './user';
+import {UserLogin} from "./model/userLogin";
 
 describe('User', () => {
   it('should create an instance', () => {
-    expect(new User()).toBeTruthy();
+    expect(new UserLogin()).toBeTruthy();
   });
 });
