@@ -1,4 +1,4 @@
 export class User {
-    email!: string;
-    password!: string;
+  email!: string;
+  password!: string;
 }
